Ok, the data model seems to be fully functional.
Note that in order to use it you will have to initialize and populate 
your own database file. The way to do that is:

django-admin.py init
django-admin.py install gappy

Note that this will create a file named gappy.db in your current 
directory, so choose wisely. The other nice part of this is that you can 
reset your test data by deleting gappy.db and repeating the two above 
steps.